Implemented 'lazy loading' for GET API calls for zone comments
The Zone comments should only load once for each zone, on initial switch to that zone. On switching to a zone, a GET request to the API pulls in the comments to a map, with the zone id as key and zone's respective comments array as its value. The zone's comments are then stored in the Front End, and if the user switches back to that zone, then no additional GET requests are performed