Created account reducer and current user received action
Connected the account reducer to the Account React container component, and the reducer holds the current user. If there is no user, then the state of the account.user is set to null and the login/signup forms will be rendered. If the user is logged in (current user not null) then a greeting message will be rendered